[goto] remove the legacy imperative exception path (#5075)
Symbolic exception lowering (remove_exceptions) now covers the whole C++ and
Python exception corpus, so make it the only exception path: it runs
unconditionally and a program it cannot lower (a concurrent program that uses
exceptions, a dynamic_cast<T&> whose std::bad_cast is unavailable) is reported
as a hard error instead of falling back.

Deletes symex_catch.cpp and the THROW/CATCH/THROW_DECL/THROW_DECL_END symex
dispatch, plus all imperative exception state (stack_catch, thrown_obj_map,
last_throw, inside_unexpected, exceptiont). merge_state_map is kept (shared with
normal call/return). report_fallback becomes report_unsupported, throwing the
ESBMC fatal-error idiom. Removes the now-meaningless --no-lower-exceptions /
--extended-try-analysis options and the ON-vs-OFF differential gate.

Fixes #5075
