+++ regression/esbmc-cpp/try_catch/lower-exceptions_bad_cast_unsupported/test.desc
@@ -1,0 +1,4 @@
+CORE
+main.cpp
+
+exception lowering: cannot lower a failing dynamic_cast
