+++ regression/esbmc-cpp/try_catch/lower-exceptions_concurrent/test.desc
@@ -1,5 +1,4 @@
 CORE
 main.cpp
---lower-exceptions --context-bound 2
 
-^VERIFICATION SUCCESSFUL$
+exception lowering: cannot lower a concurrent program
