+++ regression/esbmc-cpp/try_catch/lower-exceptions_concurrent_no_exc/test.desc
@@ -1,5 +1,5 @@
 CORE
 main.cpp
---lower-exceptions
+--no-pointer-check
 
 ^VERIFICATION SUCCESSFUL$
