--- regression/esbmc-cpp/try_catch/lower-exceptions_unsupported_warns_fail/test.desc
@@ -1,6 +1,0 @@
-CORE
-main.cpp
---lower-exceptions
-
-^VERIFICATION FAILED$
---lower-exceptions: cannot lower a concurrent program
