+++ src/esbmc/esbmc_parseoptions.cpp
@@ -2408,12 +2408,12 @@ bool esbmc_parseoptionst::process_goto_program(
       algorithm->run(goto_functions);
     }
 
-    // Lower throw/catch to symbolic guarded control flow (#5075), on by default.
-    // Run before inlining so per-call-site exception propagation is still
-    // explicit. --no-lower-exceptions selects the legacy imperative path; a
-    // program the pass cannot lower still falls back to it with a warning.
-    if (!cmdline.isset("no-lower-exceptions"))
-      remove_exceptions(goto_functions, context, ns);
+    // Lower throw/catch to symbolic guarded control flow (#5075). Run before
+    // inlining so per-call-site exception propagation is still explicit. This
+    // is now the only exception path: a program the pass cannot lower is
+    // reported as an error rather than silently miscompiled (the legacy
+    // imperative path in symex was removed once the lowered subset covered it).
+    remove_exceptions(goto_functions, context, ns);
 
     // do partial inlining
     if (!cmdline.isset("no-inlining"))
