+++ src/esbmc/options.cpp
@@ -381,16 +381,9 @@ const struct group_opt_templ all_cmd_options[] = {
      "Do not unroll bounded loops at goto level (need to enable "
      "--goto-unwind)"},
     {"slice-assumes", NULL, "Remove unused assume statements"},
-    {"extended-try-analysis",
-     NULL,
-     "Skip backward stack search for C++ throw targets"},
     {"lower-exceptions",
      NULL,
-     "Lower throw/catch to symbolic guarded control flow (default; #5075)"},
-    {"no-lower-exceptions",
-     NULL,
-     "Use the legacy imperative exception handling instead of symbolic "
-     "lowering (#5075)"},
+     "Accepted for compatibility; throw/catch lowering is always on (#5075)"},
     {"skip-bmc", NULL, "Do not perform bounded model checking"},
     {"no-cache-asserts",
      NULL,
