+++ src/goto-programs/remove_exceptions.h
@@ -13,8 +13,8 @@ class namespacet;
 /// `typeid in concrete_subtype_ids(C)` (see exception_typeidt) over a symbolic
 /// type-id, with `catch (...)` as an unconditional branch.
 ///
-/// Scope (whole-program, all-or-nothing — anything outside it leaves the entire
-/// program to the imperative symex path):
+/// Scope (whole-program, all-or-nothing — anything outside it is reported as
+/// unsupported, since this is the only exception path):
 ///   - C++ class and primitive throws; reference, value (with single-inheritance
 ///     slicing), pointer and void* catches, and catch-all;
 ///   - nested try regions with innermost-out propagation;
@@ -22,11 +22,13 @@ class namespacet;
 ///     rethrow;
 ///   - uncaught detection at main and __ESBMC_main (incl. static-init throws);
 ///   - Python try/except/raise via the same machinery.
-/// A small residual still forces fallback (std::bad_exception by value, some
-/// unusual try-block layouts); the pass warns when it declines (report_fallback).
+/// A small residual is rejected as unsupported (e.g. concurrent programs that
+/// use exceptions, --function isolated verification, std::bad_exception by
+/// value); the pass errors when it declines (report_unsupported).
 ///
-/// Run before goto_partial_inline; on by default, with --no-lower-exceptions
-/// selecting the legacy imperative dispatch as a fallback (#5075).
+/// Run before goto_partial_inline. This is the only exception path — the legacy
+/// imperative symex dispatch was removed once the lowered subset covered the
+/// corpus (#5075).
 void remove_exceptions(
   goto_functionst &goto_functions,
   contextt &context,
