+++ src/goto-symex/goto_symex.h
@@ -8,7 +8,6 @@
 #include <map>
 #include <optional>
 #include <pointer-analysis/dereference.h>
-#include <stack>
 #include <util/i2string.h>
 #include <irep2/irep2.h>
 #include <util/options.h>
@@ -546,50 +545,6 @@ protected:
   void simplify_python_builtins(expr2tc &expr);
 
   void volatile_check(expr2tc &expr);
-
-  /* Check if thrown_type in Python inherits from catch_type */
-  bool is_python_exception_subtype(
-    const irep_idt &thrown_type,
-    const irep_idt &catch_type);
-
-  /** Walk back up stack frame looking for exception handler. */
-  bool symex_throw();
-
-  /** Core throw dispatch: match throw_code against stack_catch and jump.
-   *  Called by symex_throw() and symex_throw_bad_cast(). */
-  bool symex_throw_dispatch(const expr2tc &throw_code);
-
-  /** Handle dynamic_cast<T&> failure: resolve std::bad_cast from the
-   *  namespace and dispatch the exception through the normal throw path. */
-  bool symex_throw_bad_cast();
-
-  /** Register exception handler on stack. */
-  void symex_catch();
-
-  /** Register throw handler on stack. */
-  void symex_throw_decl();
-
-  /** Update throw target. */
-  void update_throw_target(
-    goto_symex_statet::exceptiont *except [[maybe_unused]],
-    goto_programt::const_targett target,
-    const expr2tc &code,
-    bool is_ellipsis = false);
-
-  /** Check if we can rethrow an exception:
-   *  if we can then update the target.
-   *  if we can't then gives a error.
-   */
-  bool handle_rethrow(
-    const expr2tc &operand,
-    const goto_programt::instructiont &instruction);
-
-  /** Check if we can throw an exception:
-   *  if we can't then gives a error.
-   */
-  int handle_throw_decl(
-    goto_symex_statet::exceptiont *frame,
-    const irep_idt &id);
 
   /**
    *  Finalize the result of a realloc operation.
@@ -1194,33 +1149,6 @@ protected:
    *  Used to track what we should level memory-leak-assertions against when the
    *  program execution has finished */
   std::list<allocated_obj> dynamic_memory;
-
-  /* Exception Handling.
-   * This will stack the try-catch blocks, so we always know which catch
-   * we should jump.
-   */
-  typedef std::stack<goto_symex_statet::exceptiont> stack_catcht;
-
-  /** Stack of try-catch blocks. */
-  stack_catcht stack_catch;
-
-  /** Pointer to last thrown exception. */
-  goto_programt::instructiont *last_throw;
-
-  /** Backing storage for last_throw when the throw originates from the
-   *  __ESBMC_throw_bad_cast intrinsic rather than a real THROW instruction. */
-  goto_programt::instructiont bad_cast_throw;
-
-  /** Map of currently active exception targets, i.e. instructions where an
-   *  exception is going to be merged in in the future. Keys are iterators to
-   *  the instruction catching the object; values are the symbols that the
-   *  thrown piece of data has been assigned to. */
-  std::map<goto_programt::const_targett, expr2tc> thrown_obj_map;
-
-  /** Flag to indicate if we are go into the unexpected flow. */
-  bool inside_unexpected;
-  /** Store the unexpected function end */
-  irep_idt unexpected_end;
 
   /** Disable return value optimization */
   bool no_return_value_opt;
