--- src/goto-symex/symex_catch.cpp
@@ -1,560 +1,0 @@
-#include <goto-symex/goto_symex.h>
-#include <irep2/irep2.h>
-#include <util/migrate.h>
-
-void goto_symext::symex_catch()
-{
-  // there are two variants: 'push' and 'pop'
-  const goto_programt::instructiont &instruction = *cur_state->source.pc;
-
-  if (instruction.targets.empty()) // The second catch, pop from the stack
-  {
-    // Copy the exception before pop
-    goto_symex_statet::exceptiont exception = stack_catch.top();
-
-    // Pop from the stack
-    stack_catch.pop();
-
-    // Increase the program counter
-    cur_state->source.pc++;
-  }
-  else // The first catch, push it to the stack
-  {
-    goto_symex_statet::exceptiont exception;
-
-    // copy targets
-    const code_cpp_catch2t &catch_ref = to_code_cpp_catch2t(instruction.code);
-
-    assert(catch_ref.exception_list.size() == instruction.targets.size());
-
-    // Fill the map with the catch type and the target
-    unsigned i = 0;
-    for (goto_programt::targetst::const_iterator it =
-           instruction.targets.begin();
-         it != instruction.targets.end();
-         ++it, ++i)
-    {
-      exception.catch_map[catch_ref.exception_list[i]] = *it;
-      exception.catch_order[catch_ref.exception_list[i]] = i;
-    }
-
-    // Stack it
-    stack_catch.push(exception);
-
-    // Increase the program counter
-    cur_state->source.pc++;
-  }
-}
-
-bool goto_symext::is_python_exception_subtype(
-  const irep_idt &thrown_type,
-  const irep_idt &catch_type)
-{
-  std::string thrown = thrown_type.as_string();
-  std::string catch_t = catch_type.as_string();
-
-  // Exact match
-  if (thrown == catch_t)
-    return true;
-
-  // Look up the thrown exception class in the symbol table
-  const symbolt *thrown_symbol = ns.lookup("tag-" + thrown);
-  if (!thrown_symbol)
-    return false;
-
-  // Get the bases from the type metadata
-  const irept &bases = thrown_symbol->get_type().find("bases");
-
-  if (bases.is_nil())
-    return false;
-
-  const irept::subt &base_list = bases.get_sub();
-
-  // Iterate through base classes
-  for (const auto &base : base_list)
-  {
-    // The base class is stored with id() == "tag-BaseClassName"
-    std::string base_name = base.id().as_string();
-
-    // Remove "tag-" prefix to get the class name
-    if (base_name.starts_with("tag-"))
-      base_name = base_name.substr(4);
-
-    // Recursively check if this base class matches or inherits from catch_type
-    if (is_python_exception_subtype(irep_idt(base_name), catch_type))
-      return true;
-  }
-
-  return false;
-}
-
-bool goto_symext::symex_throw_dispatch(const expr2tc &throw_code)
-{
-  irep_idt catch_name = "missing";
-  const goto_programt::const_targett *catch_insn = nullptr;
-  const code_cpp_throw2t &throw_ref = to_code_cpp_throw2t(throw_code);
-  const std::vector<irep_idt> &exceptions_thrown = throw_ref.exception_list;
-
-  // We check before iterate over the throw list to save time:
-  // If there is no catch, we return an error
-  if (!stack_catch.size())
-  {
-    if (!unexpected_handler())
-    {
-      if (!terminate_handler())
-      {
-        // An un-caught exception. Error
-        const std::string &msg = "Throwing an exception of type " +
-                                 exceptions_thrown.begin()->as_string() +
-                                 " but there is not catch for it.";
-        claim(gen_false_expr(), msg);
-        return true;
-      }
-    }
-    return false;
-  }
-
-  // Get the list of catchs
-  goto_symex_statet::exceptiont *except = &stack_catch.top();
-
-  // It'll be used for catch ordering when throwing
-  // a derived object with multiple inheritance
-  unsigned old_id_number = -1, new_id_number = 0;
-
-  goto_symex_statet::call_stackt old_stack = cur_state->call_stack;
-  for (auto const &it : throw_ref.exception_list)
-  {
-    // Handle throw declarations
-    switch (handle_throw_decl(except, it))
-    {
-    case 0:
-      return true;
-      break;
-
-    case 1:
-      return false;
-      break;
-
-    case 2:
-      break;
-
-    default:
-      assert(0);
-      break;
-    }
-
-    // Search for a catch with a matching type (including base classes)
-    goto_symex_statet::exceptiont::catch_mapt::const_iterator c_it =
-      except->catch_map.find(it);
-
-    // Track which catch type was matched (might be a base class)
-    irep_idt matched_catch_type = it;
-
-    // If no exact match, check inheritance hierarchy for Python exceptions
-    if (c_it == except->catch_map.end())
-    {
-      for (const auto &catch_entry : except->catch_map)
-      {
-        if (is_python_exception_subtype(it, catch_entry.first))
-        {
-          c_it = except->catch_map.find(catch_entry.first);
-          matched_catch_type = catch_entry.first; // Track the actual catch type
-          break;
-        }
-      }
-    }
-
-    // Do we have a catch for it?
-    if (c_it != except->catch_map.end())
-    {
-      // We do!
-
-      // Get current catch number and update if needed
-      // Use matched_catch_type instead of it for the lookup
-      new_id_number = (*except->catch_order.find(matched_catch_type)).second;
-
-      if (new_id_number < old_id_number)
-      {
-        // Only restore call_stack when re-selecting a better catch handler.
-        // Skip restoration on first match (old_id_number == -1) to avoid unnecessary
-        // deep copy that causes crashes on macOS with Python exceptions.
-        if (old_id_number != (unsigned)-1)
-          cur_state->call_stack = old_stack;
-        cur_state->guard.make_true();
-
-        update_throw_target(except, c_it->second, throw_code);
-        catch_insn = &c_it->second;
-        catch_name = c_it->first;
-      }
-
-      // Save old number id
-      old_id_number = new_id_number;
-    }
-    else // We don't have a catch for it
-    {
-      // If it's a pointer, we must look for a catch(void*)
-      if (it.as_string().find("_ptr") != std::string::npos)
-      {
-        // It's a pointer!
-
-        // Do we have an void*?
-        c_it = except->catch_map.find("void_ptr");
-
-        if (c_it != except->catch_map.end())
-        {
-          // Make the jump to void*
-          update_throw_target(except, c_it->second, throw_code);
-          catch_insn = &c_it->second;
-          catch_name = c_it->first;
-        }
-      }
-      else
-      {
-        // Do we have an ellipsis?
-        c_it = except->catch_map.find("ellipsis");
-
-        if (c_it != except->catch_map.end())
-        {
-          update_throw_target(except, c_it->second, throw_code, true);
-          catch_insn = &c_it->second;
-          catch_name = c_it->first;
-        }
-      }
-    }
-  }
-
-  if (catch_insn == nullptr)
-  {
-    // No catch for type, void, or ellipsis
-    // Call terminate handler before showing error message
-    if (!terminate_handler())
-    {
-      // An un-caught exception. Error
-      const std::string &msg = "Throwing an exception of type " +
-                               exceptions_thrown.begin()->as_string() +
-                               " but there is not catch for it.";
-      claim(gen_false_expr(), msg);
-      // Ensure no further execution along this path.
-      cur_state->guard.make_false();
-    }
-
-    return false;
-  }
-
-  log_debug(
-    "symex",
-    "Caught by catch({}) at file {} line {}",
-    catch_name,
-    (*catch_insn)->location.file(),
-    (*catch_insn)->location.line());
-
-  return true;
-}
-
-bool goto_symext::symex_throw()
-{
-  const goto_programt::instructiont &instruction = *cur_state->source.pc;
-  const code_cpp_throw2t &throw_ref = to_code_cpp_throw2t(instruction.code);
-
-  if (handle_rethrow(throw_ref.operand, instruction))
-    return true;
-
-  last_throw = const_cast<goto_programt::instructiont *>(&instruction);
-
-  log_debug(
-    "symex",
-    "Exception thrown of type {} at file {} line {}",
-    throw_ref.exception_list.begin()->as_string(),
-    instruction.location.file(),
-    instruction.location.line());
-
-  return symex_throw_dispatch(instruction.code);
-}
-
-bool goto_symext::symex_throw_bad_cast()
-{
-  // Reuse the previously constructed instruction if available.
-  if (!bad_cast_throw.code)
-  {
-    // Depending on the Clang/LLVM version the <typeinfo> model's class is
-    // recorded either with the un-elaborated name ("tag-std::bad_cast") or with
-    // the elaborated one ("tag-class std::bad_cast"); newer LLVM uses the
-    // latter, which must match the name the catch clause carries. Try both.
-    const symbolt *bad_cast_sym = ns.lookup("tag-std::bad_cast");
-    if (!bad_cast_sym)
-      bad_cast_sym = ns.lookup("tag-class std::bad_cast");
-    if (!bad_cast_sym)
-    {
-      // <typeinfo> not included — emit a hard failure directly.
-      claim(
-        gen_false_expr(),
-        "dynamic_cast<T&> failed; include <typeinfo> for std::bad_cast");
-      return true;
-    }
-
-    // Build exception_list: concrete type + all direct base classes,
-    // stripping the "tag-" prefix to match the catch-side convention.
-    std::vector<irep_idt> exception_list;
-    const std::string type_id = id2string(bad_cast_sym->id);
-    exception_list.emplace_back(type_id.substr(4)); // "std::bad_cast"
-    if (bad_cast_sym->get_type().id() == "struct")
-    {
-      const struct_typet &st = to_struct_type(bad_cast_sym->get_type());
-      const exprt &bases = static_cast<const exprt &>(st.find("bases"));
-      if (bases.is_not_nil())
-        for (const auto &base : bases.get_sub())
-          exception_list.emplace_back(id2string(base.id()).substr(4));
-    }
-
-    // Build a nondet operand of bad_cast type for the thrown object.
-    type2tc bad_cast_type = migrate_symbol_type(*bad_cast_sym);
-    expr2tc nondet_op = sideeffect2tc(
-      bad_cast_type,
-      expr2tc(),
-      expr2tc(),
-      std::vector<expr2tc>(),
-      type2tc(),
-      sideeffect2t::allockind::nondet);
-    replace_nondet(nondet_op);
-
-    bad_cast_throw.make_throw();
-    bad_cast_throw.code = code_cpp_throw2tc(nondet_op, exception_list);
-    last_throw = &bad_cast_throw;
-  }
-
-  log_debug("symex", "dynamic_cast<T&> failure: throwing std::bad_cast");
-
-  return symex_throw_dispatch(bad_cast_throw.code);
-}
-
-bool goto_symext::terminate_handler()
-{
-  // We must look on the context if the user included exception lib
-  const symbolt *tmp = ns.lookup("std::terminate()");
-  bool is_included = !tmp;
-
-  // If it do, we must call the terminate function:
-  // It'll call the current function handler
-  if (!is_included)
-  {
-    codet terminate_function = to_code(tmp->get_value().op0());
-
-    // We only call it if the user replaced the default one
-    if (terminate_function.op1().identifier() == "std::default_terminate()")
-      return false;
-
-    // Call the function
-    expr2tc da_funk;
-    migrate_expr(terminate_function, da_funk);
-    symex_function_call(da_funk);
-    return true;
-  }
-
-  // If it wasn't included, we do nothing. The error message will be
-  // shown to the user as there is a throw without catch.
-  return false;
-}
-
-bool goto_symext::unexpected_handler()
-{
-  // Look if we already on the unexpected flow
-  // If true, we shouldn't call the unexpected handler again
-  if (inside_unexpected)
-    return false;
-
-  // We must look on the context if the user included exception lib
-  const symbolt *tmp = ns.lookup("c:@N@std@F@unexpected#");
-  bool is_included = !tmp;
-
-  // If it do, we must call the unexpected function:
-  // It'll call the current function handler
-  if (!is_included)
-  {
-    // We only call it if the user replaced the default one
-    const symbolt *handler = ns.lookup("c:@F@__ESBMC_unexpected");
-    if (!handler)
-      return false;
-
-    expr2tc the_call;
-    code_function_callt unexpected_function;
-    unexpected_function.function() = handler->get_value();
-    migrate_expr(unexpected_function, the_call);
-
-    // Indicate there we're inside the unexpected flow
-    inside_unexpected = true;
-
-    // Call the function
-    symex_function_call(the_call);
-    unexpected_end = handler->get_value().identifier();
-    return true;
-  }
-
-  // If it wasn't included, we do nothing. The error message will be
-  // shown to the user as there is a throw without catch.
-  return false;
-}
-
-void goto_symext::update_throw_target(
-  goto_symex_statet::exceptiont *except [[maybe_unused]],
-  goto_programt::const_targett target,
-  const expr2tc &code,
-  bool is_ellipsis)
-{
-  // Something is going to catch, therefore we need something to be caught.
-  // Assign that something to a variable and make records so that it's merged
-  // into the right place in the future.
-  assert(!is_nil_expr(code));
-  const code_cpp_throw2t &throw_insn = to_code_cpp_throw2t(code);
-
-  // Generate a name to assign this to.
-  expr2tc thrown_obj =
-    symbol2tc(throw_insn.operand->type, irep_idt("symex_throw::thrown_obj"));
-  expr2tc operand = throw_insn.operand;
-  symex_assign(code_assign2tc(thrown_obj, operand));
-
-  // Target is, as far as I can tell, always a declaration of the variable
-  // that the thrown obj ends up in, and is followed by a (blank) assignment
-  // to it. So point at the next insn.
-  //
-  // An ellipsis (catch-all) handler binds no exception variable: its target
-  // points straight at the first instruction of the handler body, which may be
-  // an ASSERT/ASSUME/GOTO whose `code` field is nil. Only inspect target->code
-  // (and record the thrown object for the bound variable) when one exists.
-  if (!is_ellipsis)
-  {
-    // Now record that value for future reference.
-    if (!is_pointer_type(target->code->type))
-      cur_state->rename(thrown_obj);
-
-    assert(is_code_decl2t(target->code));
-    target++;
-    assert(is_code_assign2t(target->code));
-    // signed int b;
-    // b = NONDET(signed int); move to this line
-
-    // Signal assignment code to fetch the thrown object and rewrite the
-    // assignment, assigning the thrown obj to the local variable.
-    thrown_obj_map.insert_or_assign(target, thrown_obj);
-  }
-
-  if (!options.get_bool_option("extended-try-analysis"))
-  {
-    // Search backwards through stack frames, looking for the frame that
-    // contains the function containing the target instruction.
-    goto_symex_statet::call_stackt::reverse_iterator i;
-    for (i = cur_state->call_stack.rbegin(); i != cur_state->call_stack.rend();
-         ++i)
-    {
-      irep_idt id = i->function_identifier.empty() ? "__ESBMC_main"
-                                                   : i->function_identifier;
-      if (id == target->function)
-      {
-        statet::merge_state_listt &merge_state_list =
-          i->merge_state_map[target];
-
-        merge_state_list.emplace_back(*cur_state);
-        cur_state->guard.make_false();
-        break;
-      }
-    }
-
-    assert(
-      (i != cur_state->call_stack.rend() ||
-       target->function == "__ESBMC_main") &&
-      "Target instruction in throw "
-      "handler not in any function frame on the stack");
-  }
-}
-
-int goto_symext::handle_throw_decl(
-  goto_symex_statet::exceptiont *except,
-  const irep_idt &id)
-{
-  // Check if we can throw the exception
-  if (except->has_throw_decl)
-  {
-    goto_symex_statet::exceptiont::throw_list_sett::const_iterator s_it =
-      except->throw_list_set.find(id);
-
-    // Is it allowed?
-    if (s_it == except->throw_list_set.end())
-    {
-      if (!unexpected_handler())
-      {
-        std::string msg =
-          std::string("Trying to throw an exception ") +
-          std::string("but it's not allowed by declaration.\n\n");
-        msg += "  Exception type: " + id.as_string();
-        msg += "\n  Allowed exceptions:";
-
-        for (const auto &s_it1 : except->throw_list_set)
-          msg += "\n   - " + std::string(s_it1.c_str());
-
-        claim(gen_false_expr(), msg);
-        return 0;
-      }
-
-      return 1;
-    }
-  }
-  return 2;
-}
-
-bool goto_symext::handle_rethrow(
-  const expr2tc &operand,
-  const goto_programt::instructiont &instruction)
-{
-  // throw without argument, we must rethrow last exception
-  if (is_nil_expr(operand))
-  {
-    if (
-      last_throw != nullptr &&
-      to_code_cpp_throw2t(last_throw->code).exception_list.size())
-    {
-      // get exception from last throw
-      std::vector<irep_idt>::const_iterator e_it =
-        to_code_cpp_throw2t(last_throw->code).exception_list.begin();
-
-      // update current state exception list
-      goto_programt::instructiont &mutable_ref =
-        const_cast<goto_programt::instructiont &>(instruction);
-      to_code_cpp_throw2t(mutable_ref.code).exception_list.push_back((*e_it));
-      to_code_cpp_throw2t(mutable_ref.code).operand =
-        to_code_cpp_throw2t(last_throw->code).operand;
-
-      return false;
-    }
-
-    const std::string &msg = "Trying to re-throw without last exception.";
-    claim(gen_false_expr(), msg);
-    return true;
-  }
-  return false;
-}
-
-void goto_symext::symex_throw_decl()
-{
-  // Check if we have a previous try-block catch
-  if (stack_catch.size())
-  {
-    const goto_programt::instructiont &instruction = *cur_state->source.pc;
-
-    // Get throw list
-    const std::vector<irep_idt> &throw_decl_list =
-      to_code_cpp_throw_decl2t(instruction.code).exception_list;
-
-    // Get to the correct try (always the last one)
-    goto_symex_statet::exceptiont *except = &stack_catch.top();
-
-    // Set the flag that this frame has throw list
-    // This is important because we can have empty throw lists
-    except->has_throw_decl = true;
-
-    // Clear before insert new types
-    except->throw_list_set.clear();
-
-    // Copy throw list to the set
-    for (const auto &i : throw_decl_list)
-      except->throw_list_set.insert(i);
-  }
-}
